Added two additional tests. The first was used to analyze why there are sometimes strange "jumps" in the SDR curve of AltLambertScalarHybridScheme. The second was used to find out why apparently the SDR can be increased by a few dB if a different constant is used in the expression for SNR^\epsilon, even though that contradicted the theoretical results.
In the meantime I have solved these puzzles, but I am leaving these functions in
here as they might be useful at a later time.

The Scheme class now always calls snr_updated() when an SNR value is set using
set_snr(), even if the new value is equal to the old value. This because it
turned out that sometimes (e.g., in test_jumps()) it was useful to run a
simulation repeatedly for the same SNR value (but with different noise samples).


git-svn-id: svn+ssh://dionysos/volume1/homes/marius/svnroot/epfldocs/work@683 2b8fe21e-8702-0410-8809-fff53b49c4e2