Makefile variable to control the use of MATLAB
I did not like that you had to switch to a completely different branch
when you did not have MATLAB available. This meant that every future
branch hat to _either_ be based on the MATLAB build or the non-MATLAB
build.

With this commit, MATLAB is used by default when you run 'make'. But
through the use of conditionals in the Makefile, you can now run

  make USEMATLAB=0

and instead of using MATLAB, the build copies already generated files
from the matlab_gen directory.